+++ apps/vr-tests/src/stories/Panel.stories.tsx
@@ -13,16 +13,18 @@ const defaultProps = {
 storiesOf('Panel', module)
   .addDecorator(FabricDecorator)
   .addDecorator(story => <Screener steps={new Screener.Steps().snapshot('default').end()}>{story()}</Screener>)
-  .add('Small left w/ close button', () => <Panel {...defaultProps} hasCloseButton type={PanelType.smallFixedNear} headerText="Small" />)
-  .add('Small fixed right w/ close button', () => (
+  .addStory('Small left w/ close button', () => (
+    <Panel {...defaultProps} hasCloseButton type={PanelType.smallFixedNear} headerText="Small" />
+  ), { rtl: true })
+  .addStory('Small fixed right w/ close button', () => (
     <Panel {...defaultProps} hasCloseButton type={PanelType.smallFixedFar} headerText="Small fixed" />
-  ))
-  .add('Small fluid right', () => <Panel {...defaultProps} type={PanelType.smallFluid} headerText="Small fluid" />)
-  .add('Medium right', () => <Panel {...defaultProps} type={PanelType.medium} headerText="Medium" />)
-  .add('Large right', () => <Panel {...defaultProps} type={PanelType.large} headerText="Large" />)
-  .add('Large fixed right', () => <Panel {...defaultProps} type={PanelType.largeFixed} headerText="Large fixed" />)
-  .add('Extra large right', () => <Panel {...defaultProps} type={PanelType.extraLarge} headerText="Extra Large" />)
-  .add('Custom', () => <Panel {...defaultProps} type={PanelType.custom} headerText="Custom" customWidth="200vw" />);
+  ), { rtl: true })
+  .addStory('Small fluid right', () => <Panel {...defaultProps} type={PanelType.smallFluid} headerText="Small fluid" />)
+  .addStory('Medium right', () => <Panel {...defaultProps} type={PanelType.medium} headerText="Medium" />, { rtl: true })
+  .addStory('Large right', () => <Panel {...defaultProps} type={PanelType.large} headerText="Large" />)
+  .addStory('Large fixed right', () => <Panel {...defaultProps} type={PanelType.largeFixed} headerText="Large fixed" />)
+  .addStory('Extra large right', () => <Panel {...defaultProps} type={PanelType.extraLarge} headerText="Extra Large" />)
+  .addStory('Custom', () => <Panel {...defaultProps} type={PanelType.custom} headerText="Custom" customWidth="200vw" />);
 
 storiesOf('Panel', module)
   .addDecorator(FabricDecorator)
@@ -37,11 +39,11 @@ storiesOf('Panel', module)
       {story()}
     </Screener>
   ))
-  .add('SearchBox and Right Panel', () => (
+  .addStory('SearchBox and Right Panel', () => (
     <div>
       <SearchBox placeholder="Search" />
       <Panel {...defaultProps} isOpen={false} type={PanelType.medium} headerClassName="" headerText={'Header'} isHiddenOnDismiss>
         {null}
       </Panel>
     </div>
-  ));
+  ), { rtl: true });