+++ scripts/generate-version-files.js
@@ -4,12 +4,14 @@ const fs = require('fs');
 const glob = require('glob');
 
 const generateOnly = process.argv.indexOf('-g') > -1;
+const rushCmd = `"${process.execPath}" "${path.resolve(__dirname, '../common/scripts/install-run-rush.js')}"`;
 
 function run(cmd) {
   return execSync(cmd, { cwd: path.resolve(__dirname, '..') }).toString();
 }
 
 let modified = [];
+let untracked = [];
 
 if (!generateOnly) {
   // Check that no uncommitted changes exist
@@ -20,12 +22,21 @@ if (!generateOnly) {
   }
 
   // Do a dry-run on all packages
-  run('rush publish -a');
+  run(`${rushCmd} publish -a`);
   status = run('git status --porcelain=1');
-  modified = status
-    .split(/\n/g)
-    .map(line => line && '"' + line.trim().split(/\s/)[1] + '"')
-    .filter(item => item);
+  status.split(/\n/g).forEach(line => {
+    if (line) {
+      const parts = line.trim().split(/\s/);
+
+      if (parts[0] === '??') {
+        // untracked files at this point would be things like CHANGELOG files for a brand new project
+        untracked.push(parts[1]);
+      } else {
+        // modified files include package.json, generated CHANGELOG files from rush publish dry run
+        modified.push('"' + parts[1] + '"');
+      }
+    }
+  });
 }
 
 const packageJsons = glob.sync('+(packages|apps)/*/package.json');
@@ -58,6 +69,9 @@ setVersion('${packageJson.name}', '${packageJson.version}');`
 
 if (!generateOnly) {
   // Undo the dry-run changes, preserve the version file changes
+  console.log(`remove untracked ${untracked.join(' ')}`);
+  untracked.forEach(f => fs.unlinkSync(f));
+
   console.log(`reset ${modified.join(' ')}`);
   run(`git checkout ${modified.join(' ')}`);
 }