+++ packages/dashboard/CHANGELOG.md
@@ -1,6 +1,14 @@
 # Change Log - @uifabric/dashboard
 
-This log was last generated on Thu, 20 Sep 2018 12:25:33 GMT and should not be manually modified.
+This log was last generated on Fri, 21 Sep 2018 14:25:46 GMT and should not be manually modified.
+
+## 0.25.4
+Fri, 21 Sep 2018 14:25:46 GMT
+
+### Patches
+
+- Adding a version stamp file
+- Remove drag on menu button on card frame and link button 
 
 ## 0.25.3
 Thu, 20 Sep 2018 12:25:33 GMT